pyproject.toml
[tool.poetry]
name = "llm-as-evaluator-service"
version = "1.0.0"
description = "Service to run evaluations for SFT and RLHF tasks using LLMs."
authors = ["Alexei Vasilkov <alexei.v@turing.com>", "Navaneethan <navaneethan.ramasamy@turing.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "*"
gevent = "*"
langchain = "*"
redis = "*"
asyncmy = "*"
sqlalchemy = { version = "*", extras = ["asyncio"] }
celery = "*"
tqdm = "*"
alembic = "*"
aiosignal = "*"
langgraph = "*"
fastapi = "*"
itsdangerous = "*"
tiktoken = "*"
python-dotenv = "*"
PyMySQL = "*"
json_logging = "*"
httpx = "*"
passlib = "*"
pydantic = { version = "*", extras = ["email"] }
python-jose = "*"
nbformat = "*"
fuzzywuzzy = "*"
langchain-openai = "*"
langchain-community = "*"
langchain-anthropic = "*"
langchain-fireworks = "*"
langchain-groq = "*"
google-api-core = "*"
google-api-python-client = "*"
google-auth = "*"
google-auth-httplib2 = "*"
google-auth-oauthlib = "*"
google-cloud = "*"
google-cloud-core = "*"
google-cloud-pubsub = "*"
google-cloud-storage = "*"
google-crc32c = "*"
google-resumable-media = "*"
googleapis-common-protos = "*"
opentelemetry-api = "*"
opentelemetry-sdk = "*"
opentelemetry-instrumentation-celery = "*"
poetry = "*"
requests = "*"
Authlib = "*"
Deprecated = "*"
Mako = "*"
PyJWT = "*"
PyYAML = "*"
Pygments = "*"
StrEnum = "*"
autoflake = "*"
bcrypt = "*"
dirtyjson = "*"
email_validator = "*"
et-xmlfile = "*"
fonttools = "*"
langchain-experimental = "*"
typing_extensions = "*"
langchain-google-genai = "*"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytz = "^2024.2"
python-dateutil = "^2.8.2"
mistralai = "^1.2.2"
copyleaks = "^4.2.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
mypy = "^1.10.1"
pytest-asyncio = "^0.23.8"
aiosqlite = "^0.20.0"
ipykernel = "^6.29.5"
pylance = "*"
nltk = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
